# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Citrix Backend
# during dev, should point to your local machine
host: localhost:3000
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /api/v1/proxies:
    # binds a127 app logic to a route
    x-swagger-router-controller: proxies
    get:
      description: Returns 'All Proxies' to the caller
      # used as the method name of the controller
      operationId: proxies
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ProxiesResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Add new Proxy
      # used as the method name of the controller
      operationId: proxies_add
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ProxiesResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  ProxiesResponse:
    required:
      - status
      - result
    properties:
      status:
        type: number
      result:
        type: object
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string